feature/add-different-fetches, выделил провайдера запросов в отдельный файл
